python/src/auth/requirements.txt
flask==2.0.2
astroid==2.9.3
click==8.0.3
Flask-MySQLdb==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
parso==0.8.3
platformdirs==2.4.1
Werkzeug==2.0.2
PyJWT==2.3.0
mysqlclient==2.1.0
toml==0.10.2
pylint==2.12.2
wrapt==1.13.3